requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq
langchain-huggingface
flask
python-dotenv
sentence-transformers
